5 commit

+++ src/main/java/tn/esprit/bahasallami/repository/CampanyRepository.java
@@ -3,5 +3,6 @@ package tn.esprit.bahasallami.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.bahasallami.entities.Campany;
 
-public interface CampanyRepository extends JpaRepository<Campany, Integer> {
+public interface CampanyRepository extends JpaRepository<Campany, Long> {
 }
+
