+++ src/main/java/tn/esprit/bahasallami/repository/Donationrepository.java
@@ -9,10 +9,11 @@ import tn.esprit.bahasallami.entities.DonationType;
 import java.util.Date;
 import java.util.List;
 
-public interface Donationrepository extends JpaRepository<Donation, Integer>{
+public interface Donationrepository extends JpaRepository<Donation, Long>{
     //get donations by type
     public List<Donation> findByDonType(DonationType type);
-    @Query("SELECT SUM(d.donAmount) FROM Donation d WHERE d.donDate BETWEEN :startDate AND :endDate")
-    Float sumDonationsBetweenDates(@Param("startDate") Date startDate, @Param("endDate") Date endDate);
+
+   @Query("SELECT SUM(d.donAmount) FROM Donation d WHERE d.donDate BETWEEN :startDate AND :endDate")
+   Float sumDonationsBetweenDates(@Param("startDate") Date startDate, @Param("endDate") Date endDate);
 
 }
