+++ src/main/java/tn/esprit/bahasallami/repository/EmployeRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.bahasallami.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.bahasallami.entities.Employe;
 
-public interface EmployeRepository extends JpaRepository<Employe, Integer> {
+public interface EmployeRepository extends JpaRepository<Employe, Long> {
 }
